+++ phpmyfaq/admin/export.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * JSON, HTML5 and PDF export
+ * JSON, and PDF export
  *
  * This Source Code Form is subject to the terms of the Mozilla Public License,
  * v. 2.0. If a copy of the MPL was not distributed with this file, You can